change list management
