+++ src/main/java/uni/climatemonitor/data/ClimateParametersFileHandler.java
@@ -1,0 +1,15 @@
+package uni.climatemonitor.data;
+
+public class ClimateParametersFileHandler extends FileHandler {
+    public ClimateParametersFileHandler(String fileName){
+        super(fileName);
+    }
+
+    /**
+     * Read the climate params yaml file
+     */
+    @Override
+    public void readFile() {
+
+    }
+}
