+++ src/main/java/uni/climatemonitor/data/Location.java
@@ -1,4 +1,38 @@
 package uni.climatemonitor.data;
+/**
+ * Coordinates handling with some useful methods
+ */
+class Coordinates {
+    private double latitude;
+    private double longitude;
+
+    /**
+     * Constructor
+     * @param lat
+     * @param lon
+     */
+    public Coordinates(double lat, double lon){
+        this.latitude = lat;
+        this.longitude = lon;
+    }
+
+    /**
+     * Get a beautiful representation for coordinates in the form:
+     *      N... E...
+     */
+    public String toString(){
+        String lat = String.valueOf(Math.abs(latitude));
+        String lon = String.valueOf(Math.abs(longitude));
+        String NS = "N";
+        String EW = "E";
+        if (latitude < 0){ NS = "S"; }
+        if (longitude < 0){ EW = "W"; }
+        String out = lat + "° " + NS + " " + lon + "° " + EW;
+        return out;
+    }
+
+}
+
 
 /**
  * This class implements a container for Locations with associated
@@ -10,7 +44,6 @@ public class Location {
     private final String asciiName;
     private final String state;
     private final Coordinates coordinates;
-    private final String representation;
 
     /**
      * The constructor expects:
@@ -25,7 +58,6 @@ public class Location {
         state = locationInfo[4] + " (" + locationInfo[3] + ")";
         double[] coords = unpackCoordinateString(locationInfo[5]);
         coordinates = new Coordinates(coords[0], coords[1]);
-        representation = createRepresentation();
     }
 
     private double[] unpackCoordinateString(String c) {
@@ -36,16 +68,14 @@ public class Location {
         return out;
     }
 
-    private String createRepresentation() {
+    /**
+     * override for toString
+     * @return a representation of the object
+     */
+    @Override
+    public String toString() {
         String out = asciiName + ", " + state + ", " + coordinates.toString();
         return out;
-    }
-
-    /**
-     * Getter for representation
-     */
-    public String getRepresentation() {
-        return representation;
     }
 
     /**
