+++ src/main/java/uni/climatemonitor/data/LocationsFileHandler.java
@@ -11,41 +11,6 @@ import java.util.HashMap;
 import java.util.Map;
 
 /**
- * Coordinates handling with some useful methods
- */
-class Coordinates {
-    private double latitude;
-    private double longitude;
-
-    /**
-     * Constructor
-     * @param lat
-     * @param lon
-     */
-    public Coordinates(double lat, double lon){
-        this.latitude = lat;
-        this.longitude = lon;
-    }
-
-    /**
-     * Get a beautiful representation for coordinates in the form:
-     *      N... E...
-     */
-    public String toString(){
-        String lat = String.valueOf(Math.abs(latitude));
-        String lon = String.valueOf(Math.abs(longitude));
-        String NS = "N";
-        String EW = "E";
-        if (latitude < 0){ NS = "S"; }
-        if (longitude < 0){ EW = "W"; }
-        String out = lat + "° " + NS + " " + lon + "° " + EW;
-        return out;
-    }
-
-}
-
-
-/**
  * Implements methods to handle monitoring_coordinates.data
  */
 public class LocationsFileHandler extends FileHandler {
